+++ Magic/CHANGELOG.md
@@ -18,7 +18,8 @@
  - Add "direction" and "use_target" parameters to CheckBlock action, for easily checking the block under your feet
  - Fix some consistency issues with the "target_self" parameter
  - Add "target_mount" parameter (false by default) so spells won't kill your horse
- - Fix Dragonball not earning SP on h
+ - Fix Dragonball not earning SP on hit
+ - Add more detailed support for potion effect lists in PotionEffect action
 
 # 7.9.1